feat: added the hardcoded components
