+++ app/restaurant/[slug]/components/RestaurantTitle.tsx
@@ -1,0 +1,9 @@
+const RestaurantTitle = () => {
+  return (
+    <div className="mt-4 border-b pb-6">
+      <h2 className="font-bold text-6xl">Milesstone Grill</h2>
+    </div>
+  );
+};
+
+export default RestaurantTitle;
