+++ app/restaurant/[slug]/components/Reviews.tsx
@@ -1,0 +1,16 @@
+import ReviewCard from "./ReviewCard";
+
+const Reviews = () => {
+  return (
+    <div>
+      <h3 className="font-bold text-3xl mt-10 mb-7 borber-b pb-5">
+        What 100 people are saying
+      </h3>
+      <div>
+        <ReviewCard />
+      </div>
+    </div>
+  );
+};
+
+export default Reviews;
